Refactor ActiveRecord::Preloader associations
Removed the subclasses for Collection and Singular associations:
they are only different by one line of code which doesn't deserve a
hierarchy of 6 classes.

Introduced only ThroughAssociation and DirectAssociation inherited from
Association and implementing abstract #run method.
