--- activerecord/lib/active_record/associations/preloader/belongs_to.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module ActiveRecord
-  module Associations
-    class Preloader
-      class BelongsTo < SingularAssociation #:nodoc:
-      end
-    end
-  end
-end
