--- activerecord/lib/active_record/associations/preloader/has_many.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module ActiveRecord
-  module Associations
-    class Preloader
-      class HasMany < CollectionAssociation #:nodoc:
-      end
-    end
-  end
-end
