--- activerecord/lib/active_record/associations/preloader/has_one.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module ActiveRecord
-  module Associations
-    class Preloader
-      class HasOne < SingularAssociation #:nodoc:
-      end
-    end
-  end
-end
