--- activerecord/lib/active_record/associations/preloader/has_one_through.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module ActiveRecord
-  module Associations
-    class Preloader
-      class HasOneThrough < SingularAssociation #:nodoc:
-        include ThroughAssociation
-      end
-    end
-  end
-end
